# api/openapi/expensor.openapi.yaml
basePath: /api
definitions:
...
  api.AnnualHeatmapResponse:
    properties:
      buckets:
        items:
          $ref: '#/definitions/api.DailyBucketResponse'
        type: array
      year:
        example: 2026
        type: integer
    type: object
...
  api.CategoryMonthlyEntryResponse:
    properties:
      current:
        example: 1200.5
        type: number
      prior:
        example: 950.25
        type: number
    type: object
...
  api.ChartDataResponse:
    properties:
      by_bank:
        additionalProperties:
          type: number
        type: object
      by_bucket:
        additionalProperties:
          type: number
        type: object
      by_category:
        additionalProperties:
          type: number
        type: object
      by_category_monthly:
        additionalProperties:
          $ref: '#/definitions/api.CategoryMonthlyEntryResponse'
        type: object
      by_label:
        additionalProperties:
          type: number
        type: object
      by_source:
        additionalProperties:
          type: number
        type: object
      by_source_type:
        additionalProperties:
          type: number
        type: object
      daily_spend:
        items:
          $ref: '#/definitions/api.TimeBucketResponse'
        type: array
      monthly_spend:
        items:
          $ref: '#/definitions/api.TimeBucketResponse'
        type: array
    type: object
...
  api.DailyBucketResponse:
    properties:
      amount:
        example: 249.5
        type: number
      count:
        example: 3
        type: integer
      date:
        type: string
    type: object
  api.DashboardDataResponse:
    properties:
      all_time:
        $ref: '#/definitions/api.DashboardSectionResponse'
      current_month:
        $ref: '#/definitions/api.DashboardSectionResponse'
    type: object
  api.DashboardSectionResponse:
    properties:
      charts:
        $ref: '#/definitions/api.ChartDataResponse'
      label:
        example: Current Month
        type: string
      stats:
        $ref: '#/definitions/api.StatsResponse'
    type: object
  api.DayOfMonthBucketResponse:
    properties:
      amount:
        example: 249.5
        type: number
      count:
        example: 3
        type: integer
      day:
        example: 15
        type: integer
    type: object
  api.ErrorResponse:
    properties:
      error:
        example: internal server error
        type: string
    type: object
...
  api.HeatmapDataResponse:
    properties:
      by_day_of_month:
        items:
          $ref: '#/definitions/api.DayOfMonthBucketResponse'
        type: array
      by_weekday_hour:
        items:
          $ref: '#/definitions/api.WeekdayHourBucketResponse'
        type: array
    type: object
...
  api.StatsResponse:
    properties:
      base_currency:
        example: INR
        type: string
      total_base:
        type: number
      total_by_category:
        additionalProperties:
          type: number
        type: object
      total_category_count:
        additionalProperties:
          type: integer
        type: object
      total_count:
        type: integer
    type: object
  api.StatusOnlyResponse:
    properties:
      status:
        example: ok
        type: string
    type: object
...
  api.SyncStatusResponse:
    properties:
      entries_updated:
        type: integer
      error:
        type: string
        x-nullable: true
      last_synced_at:
        type: string
        x-nullable: true
    type: object
...
  api.TimeBucketResponse:
    properties:
      amount:
        example: 249.5
        type: number
      count:
        example: 3
        type: integer
      period:
        example: 2026-05
        type: string
    type: object
...
  api.WeekdayHourBucketResponse:
    properties:
      amount:
        example: 249.5
        type: number
      count:
        example: 3
        type: integer
      hour:
        example: 14
        type: integer
      weekday:
        example: 1
        type: integer
    type: object
...
info:
  contact: {}
  description: This is the generated API contract for the Expensor backend.
  title: Expensor API
  version: dev
paths:
...
  /config/sync:
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.StatusOnlyResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Trigger community content sync
      tags:
      - Config
  /config/sync/status:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SyncStatusResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Get community sync status
      tags:
      - Config
...
  /stats/charts:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ChartDataResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Get chart data
      tags:
      - Stats
  /stats/dashboard:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.DashboardDataResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Get dashboard data
      tags:
      - Stats
  /stats/heatmap:
    get:
      parameters:
      - description: RFC3339 start timestamp
        example: "2026-05-01T00:00:00Z"
        in: query
        name: from
        type: string
      - description: RFC3339 end timestamp
        example: "2026-05-31T23:59:59Z"
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.HeatmapDataResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Get spending heatmap
      tags:
      - Stats
  /stats/heatmap/annual:
    get:
      parameters:
      - description: Calendar year
        example: 2026
        in: query
        name: year
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.AnnualHeatmapResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Get annual spending heatmap
      tags:
      - Stats
...
schemes:
- http
- https
swagger: "2.0"