# backend/internal/api/handlers_stats.go
package api

import (
	"fmt"
	"net/http"
	"strconv"
	"time"
)

// GetChartData handles GET /api/stats/charts.
// @Summary Get chart data
// @Tags Stats
// @Produce json
// @Success 200 {object} ChartDataResponse
// @Failure 500 {object} ErrorResponse
// @Failure 503 {object} ErrorResponse
// @Router /stats/charts [get]
func (h *Handlers) GetChartData(w http.ResponseWriter, r *http.Request) {
	cd, err := h.analyticsStore.GetChartData(r.Context())
	if err != nil {
		h.logger.Error("get chart data", "error", err)
		writeError(w, http.StatusInternalServerError, "failed to fetch chart data")
		return
	}
	writeJSON(w, http.StatusOK, cd)
}

// GetDashboardData handles GET /api/stats/dashboard.
// @Summary Get dashboard data
// @Tags Stats
// @Produce json
// @Success 200 {object} DashboardDataResponse
// @Failure 500 {object} ErrorResponse
// @Failure 503 {object} ErrorResponse
// @Router /stats/dashboard [get]
func (h *Handlers) GetDashboardData(w http.ResponseWriter, r *http.Request) {
	data, err := h.analyticsStore.GetDashboardData(r.Context())
	if err != nil {
		h.logger.Error("get dashboard data", "error", err)
		writeError(w, http.StatusInternalServerError, "failed to fetch dashboard data")
		return
	}

	writeJSON(w, http.StatusOK, data)
}

// GetHeatmap handles GET /api/stats/heatmap.
// Optional query params: from=<RFC3339>, to=<RFC3339> (both or neither).
// Returns 400 if either param is present but malformed.
// @Summary Get spending heatmap
// @Tags Stats
// @Produce json
// @Param from query string false "RFC3339 start timestamp" example(2026-05-01T00:00:00Z)
// @Param to query string false "RFC3339 end timestamp" example(2026-05-31T23:59:59Z)
// @Success 200 {object} HeatmapDataResponse
// @Failure 400 {object} ErrorResponse
// @Failure 500 {object} ErrorResponse
// @Failure 503 {object} ErrorResponse
// @Router /stats/heatmap [get]
func (h *Handlers) GetHeatmap(w http.ResponseWriter, r *http.Request) {
	from, to, err := parseHeatmapRange(r)
	if err != nil {
		writeError(w, http.StatusBadRequest, err.Error())
		return
	}

	data, err := h.analyticsStore.GetSpendingHeatmap(r.Context(), from, to)
	if err != nil {
		h.logger.Error("get heatmap", "error", err)
		writeError(w, http.StatusInternalServerError, "failed to fetch heatmap data")
		return
	}
	writeJSON(w, http.StatusOK, data)
}

// GetAnnualHeatmap handles GET /api/stats/heatmap/annual?year=YYYY.
// Returns per-day transaction totals for the requested calendar year.
// Defaults to the current year when ?year is absent or invalid.
// @Summary Get annual spending heatmap
// @Tags Stats
// @Produce json
// @Param year query int false "Calendar year" example(2026)
// @Success 200 {object} AnnualHeatmapResponse
// @Failure 500 {object} ErrorResponse
// @Failure 503 {object} ErrorResponse
// @Router /stats/heatmap/annual [get]
func (h *Handlers) GetAnnualHeatmap(w http.ResponseWriter, r *http.Request) {
	yearStr := r.URL.Query().Get("year")
	year, err := strconv.Atoi(yearStr)
	if err != nil || year < 1 {
		year = time.Now().Year()
	}

	buckets, err := h.analyticsStore.GetAnnualSpend(r.Context(), year)
	if err != nil {
		h.logger.Error("get annual heatmap", "error", err, "year", year)
		writeError(w, http.StatusInternalServerError, "failed to fetch annual heatmap data")
		return
	}

	writeJSON(w, http.StatusOK, map[string]any{
		"year":    year,
		"buckets": buckets,
	})
}

// parseHeatmapRange parses optional ?from= and ?to= RFC3339 query parameters.
// Returns nil, nil when neither is provided. Returns an error if either is
// present but cannot be parsed as RFC3339.
func parseHeatmapRange(r *http.Request) (from, to *time.Time, err error) {
	if v := r.URL.Query().Get("from"); v != "" {
		t, parseErr := time.Parse(time.RFC3339, v)
		if parseErr != nil {
			return nil, nil, fmt.Errorf("invalid 'from' param: must be RFC3339 (e.g. 2026-04-01T00:00:00Z)")
		}
		from = &t
	}
	if v := r.URL.Query().Get("to"); v != "" {
		t, parseErr := time.Parse(time.RFC3339, v)
		if parseErr != nil {
			return nil, nil, fmt.Errorf("invalid 'to' param: must be RFC3339 (e.g. 2026-04-30T23:59:59Z)")
		}
		to = &t
	}
	return from, to, nil
}